ath6kl: add support for Samsung 3.0.70.X firmware
On Samsung msm7x27 devices, the latest ath6kl hw2.0 firmware
3.1.1.149 exhibits instability (poor rx transmission, frequent
re-associations to the AP). The only solution for this problem
is to use firmware provided by Samsung, but this firmware was
intended to be used by the older SDK3.0 (olca-3.0) driver which
used different fw address mappings.

Although the older firmware functions perfectly for regular
station mode, ad-hoc mode causes the driver to freeze. For this
reason, a new module parameter (samsung_firmware) has been
added in this change to allow selection of the fimware type.
Set samsung_firmware to 1 to enable loading of Samsung firmware.

Samsung firmware images (except for board data) should be renamed
with "samsung_" prefix to allow coexistence with newer firmware.